Move register delegate modal to src
