+++ src/modules/dpos/validator/components/registerDelegateForm/form.css
