+++ src/modules/dpos/validator/components/registerDelegateForm/form.test.js
