+++ src/modules/dpos/validator/components/registerDelegateForm/index.js
