+++ src/modules/dpos/validator/components/registerDelegateStatus/index.js
