+++ src/modules/dpos/validator/components/registerDelegateStatus/status.js
