+++ src/modules/dpos/validator/components/registerDelegateStatus/status.test.js
