+++ src/modules/dpos/validator/components/registerDelegateStatus/statusMessages.js
