+++ src/modules/dpos/validator/components/registerDelegateSummary/index.js
