+++ src/modules/dpos/validator/components/registerDelegateSummary/summary.css
