+++ src/modules/dpos/validator/components/registerDelegateSummary/summary.js
