+++ src/modules/dpos/validator/components/registerDelegateSummary/summary.test.js
