+++ src/modules/dpos/validator/components/registerDelegateView/registerDelegate.css
@@ -1,4 +1,4 @@
-@import '../../../../../setup/react/app/mixins.css';
+@import '../../../../../../setup/react/app/mixins.css';
 
 .wrapper {
   max-width: 720px;
