+++ src/modules/dpos/validator/manager/registerDelegateManager.js
@@ -1,6 +1,6 @@
 /* istanbul ignore file */
 import { withRouter } from 'react-router';
 import { withTranslation } from 'react-i18next';
-import registerDelegate from './registerDelegate';
+import registerDelegate from '../components/registerDelegateView';
 
 export default withRouter(withTranslation()(registerDelegate));
